Move useAuth hook to separate file to fix fast refresh warning
